Move find_subsegment and _search_entity functions to tests/util.py
