+++ tests/ext/flask_sqlalchemy/test_query.py
@@ -1,11 +1,10 @@
 from __future__ import absolute_import
-import json
-import jsonpickle
 import pytest
 from aws_xray_sdk.core import xray_recorder
 from aws_xray_sdk.core.context import Context
 from aws_xray_sdk.ext.flask_sqlalchemy.query import XRayFlaskSqlAlchemy
 from flask import Flask
+from ...util import find_subsegment
 
 
 app = Flask(__name__)
@@ -21,33 +20,6 @@ class User(db.Model):
     name = db.Column(db.String(255), nullable=False, unique=True)
     fullname = db.Column(db.String(255), nullable=False)
     password = db.Column(db.String(255), nullable=False)
-
-
-def _search_entity(entity, name):
-    """Helper function to that recursivly looks at subentities
-    Returns a serialized entity that matches the name given or None"""
-    if 'name' in entity:
-        my_name = entity['name']
-        if my_name == name:
-            return entity
-        else:
-            if "subsegments" in entity:
-                for s in entity['subsegments']:
-                    result = _search_entity(s, name)
-                    if result is not None:
-                        return result
-    return None
-
-
-def find_sub(segment, name):
-    """Helper function to find a subsegment by name in the entity tree"""
-    segment = jsonpickle.encode(segment, unpicklable=False)
-    segment = json.loads(segment)
-    for entity in segment['subsegments']:
-        result = _search_entity(entity, name)
-        if result is not None:
-            return result
-    return None
 
 
 @pytest.fixture()
@@ -67,7 +39,7 @@ def test_all(capsys, session):
     Verify that we capture trace of query and return the SQL as metdata"""
     # with capsys.disabled():
     User.query.all()
-    subsegment = find_sub(xray_recorder.current_segment(), 'sqlalchemy.orm.query.all')
+    subsegment = find_subsegment(xray_recorder.current_segment(), 'sqlalchemy.orm.query.all')
     assert subsegment['name'] == 'sqlalchemy.orm.query.all'
     assert subsegment['metadata']['default']['sql']
 
@@ -78,5 +50,5 @@ def test_add(capsys, session):
     # with capsys.disabled():
     john = User(name='John', fullname="John Doe", password="password")
     db.session.add(john)
-    subsegment = find_sub(xray_recorder.current_segment(), 'sqlalchemy.orm.session.add')
+    subsegment = find_subsegment(xray_recorder.current_segment(), 'sqlalchemy.orm.session.add')
     assert subsegment['name'] == 'sqlalchemy.orm.session.add'
