+++ tests/ext/sqlalchemy/test_query.py
@@ -1,13 +1,11 @@
 from __future__ import absolute_import
 import pytest
-import json
-import jsonpickle
 from aws_xray_sdk.core import xray_recorder
 from aws_xray_sdk.core.context import Context
 from aws_xray_sdk.ext.sqlalchemy.query import XRaySessionMaker
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy import create_engine, Column, Integer, String
-# from sqlalchemy.orm import sessionmaker
+from ...util import find_subsegment
 
 
 Base = declarative_base()
@@ -20,33 +18,6 @@ class User(Base):
         name = Column(String)
         fullname = Column(String)
         password = Column(String)
-
-
-def _search_entity(entity, name):
-    """Helper function to that recursivly looks at subentities
-    Returns a serialized entity that matches the name given or None"""
-    if 'name' in entity:
-        my_name = entity['name']
-        if my_name == name:
-            return entity
-        else:
-            if "subsegments" in entity:
-                for s in entity['subsegments']:
-                    result = _search_entity(s, name)
-                    if result is not None:
-                        return result
-    return None
-
-
-def find_sub(segment, name):
-    """Helper function to find a subsegment by name in the entity tree"""
-    segment = jsonpickle.encode(segment, unpicklable=False)
-    segment = json.loads(segment)
-    for entity in segment['subsegments']:
-        result = _search_entity(entity, name)
-        if result is not None:
-            return result
-    return None
 
 
 @pytest.fixture()
@@ -69,7 +40,7 @@ def test_all(capsys, session):
     Verify we run the query and return the SQL as metdata"""
     # with capsys.disabled():
     session.query(User).all()
-    sub = find_sub(xray_recorder.current_segment(), 'sqlalchemy.orm.query.all')
+    sub = find_subsegment(xray_recorder.current_segment(), 'sqlalchemy.orm.query.all')
     assert sub['name'] == 'sqlalchemy.orm.query.all'
     assert sub['metadata']['default']['sql']
 
@@ -80,5 +51,5 @@ def test_add(capsys, session):
     # with capsys.disabled():
     john = User(name='John', fullname="John Doe", password="password")
     session.add(john)
-    sub = find_sub(xray_recorder.current_segment(), 'sqlalchemy.orm.session.add')
+    sub = find_subsegment(xray_recorder.current_segment(), 'sqlalchemy.orm.session.add')
     assert sub['name'] == 'sqlalchemy.orm.session.add'
