+++ tests/util.py
@@ -1,5 +1,6 @@
 import json
 import threading
+import jsonpickle
 
 from aws_xray_sdk.core.recorder import AWSXRayRecorder
 from aws_xray_sdk.core.emitters.udp_emitter import UDPEmitter
@@ -43,3 +44,30 @@ def entity_to_dict(trace_entity):
 
     raw = json.loads(trace_entity.serialize())
     return raw
+
+
+def _search_entity(entity, name):
+    """Helper function to that recursivly looks at subentities
+    Returns a serialized entity that matches the name given or None"""
+    if 'name' in entity:
+        my_name = entity['name']
+        if my_name == name:
+            return entity
+        else:
+            if "subsegments" in entity:
+                for s in entity['subsegments']:
+                    result = _search_entity(s, name)
+                    if result is not None:
+                        return result
+    return None
+
+
+def find_subsegment(segment, name):
+    """Helper function to find a subsegment by name in the entity tree"""
+    segment = jsonpickle.encode(segment, unpicklable=False)
+    segment = json.loads(segment)
+    for entity in segment['subsegments']:
+        result = _search_entity(entity, name)
+        if result is not None:
+            return result
+    return None
